requirements.txt
Django==5.0.6
django-summernote==0.8.20.0
django-ninja==1.2.1
django-imagekit==5.0.0
django-filter==24.2
python-decouple==3.8